No changes

+++ CSV Analyzer Pro/CSV Analyzer Pro/Form1.cs
@@ -31,6 +31,8 @@ namespace CSV_Analyzer_Pro{
         bool _exiting = false;
 
         string Filters = "csv files (*.csv)|*.csv";
+
+        List<TabExtraInfo> tabMetadataList = new List<TabExtraInfo>();
         #endregion
 
         #region Initializing
@@ -47,6 +49,10 @@ namespace CSV_Analyzer_Pro{
 
             //Enable Progress Reorting
             worker.WorkerReportsProgress = true;
+            
+            //Create tab metadata for Welcome page
+            TabExtraInfo welcomePage = new TabExtraInfo(0, "N/A");
+            tabMetadataList.Insert(0, welcomePage);
         }
 
         private void Form1_Load(object sender, EventArgs e) {
@@ -70,7 +76,9 @@ namespace CSV_Analyzer_Pro{
 
         private void saveToolStripMenuItem_Click(object sender, EventArgs e) {
             if (!IsWelcomePage()) {
-                Thread th = new Thread(() => Save(path));
+                TabExtraInfo tabInfo;
+                tabInfo = tabMetadataList.ElementAt(tabControl1.SelectedIndex);
+                Thread th = new Thread(() => Save(tabInfo.GetAssocaitedFileName()));
                 th.Start();
             }
         }
@@ -79,6 +87,10 @@ namespace CSV_Analyzer_Pro{
             if (!IsWelcomePage()) {
                 SaveAs();
             }
+        }
+
+        private void saveAllToolStripMenuItem_Click(object sender, EventArgs e) {
+            SaveAll();
         }
 
         private void addNewRowToolStripMenuItem_Click(object sender, EventArgs e) {
@@ -154,10 +166,18 @@ namespace CSV_Analyzer_Pro{
 
         #region EventHandlers
         private void OnTabMouseUp(object sender, MouseEventArgs e) {
+            TabExtraInfo tabInfo;
+            tabInfo = tabMetadataList.ElementAt(tabControl1.SelectedIndex);
+            bool hasUnsavedChanges = tabInfo.QueryHasUnsavedChanges();
             if (e.Button == MouseButtons.Right) {
-                DialogResult result = MessageBox.Show("Do you really wanna delete this page? All unsaved data will be lost!", "Confirmation", MessageBoxButtons.YesNoCancel);
-                if (result == DialogResult.Yes) {
-                    DeleteTab();
+
+                if (hasUnsavedChanges) {
+                    PromptUnsavedChangesCloseTab(tabInfo);
+                } else { 
+                    DialogResult result = MessageBox.Show("Do you really want to close this file?", "Confirmation", MessageBoxButtons.YesNoCancel);
+                    if (result == DialogResult.Yes) {
+                        DeleteTab();
+                    }
                 }
             }
         }
@@ -185,6 +205,16 @@ namespace CSV_Analyzer_Pro{
                 e.Style.BackColor = Color.LightCyan;
             else
                 e.Style.BackColor = Color.GhostWhite;
+        }
+
+        private void GridCellChanged(object sender, GridCellsChangedEventArgs e) {
+            int index = tabControl1.SelectedIndex;
+            TabExtraInfo tabInfo;
+            tabInfo = tabMetadataList.ElementAt(index);
+            if (!tabInfo.QueryHasUnsavedChanges()) {
+                tabInfo.SetHasUnsavedChanges(true);
+                tabControl1.SelectedTab.Text = "*" + tabControl1.SelectedTab.Text; // Add asterisk to denote unsaved changes
+            }
         }
         #endregion
 
@@ -214,6 +244,18 @@ namespace CSV_Analyzer_Pro{
             if (tabControl1.SelectedIndex == 0) {
                 NewWindow();
             }
+
+            bool safeToOpen = true;
+            TabExtraInfo tabInfo;
+            tabInfo = tabMetadataList.ElementAt(tabControl1.SelectedIndex);
+            if (tabInfo.QueryHasUnsavedChanges()) {
+                safeToOpen = PromptUnsavedChangesOpenTab(tabInfo); // Checking for unsaved changes since opening a new file deletes the current one
+            }
+
+            if (!safeToOpen) {
+                return;
+            }
+
             OpenFileDialog csvSearch = new OpenFileDialog();
             csvSearch.Filter = Filters;
             csvSearch.FilterIndex = 1;
@@ -265,6 +307,10 @@ namespace CSV_Analyzer_Pro{
                 //Bind data source
                 dbg.DataSource = ds.Tables[index.ToString()];
             }
+            TabExtraInfo tabInfo;
+            tabInfo = tabMetadataList.ElementAt(tabControl1.SelectedIndex);
+            tabInfo.SetAssociatedFileName(tabControl1.SelectedTab.Text);
+            tabInfo.SetHasUnsavedChanges(false);
         }
         #endregion
 
@@ -307,6 +353,7 @@ namespace CSV_Analyzer_Pro{
             dbg.BaseStylesMap["Row Header"].StyleInfo.CellType = "Header";
 
             dbg.Model.QueryCellInfo += new Syncfusion.Windows.Forms.Grid.GridQueryCellInfoEventHandler(Model_QueryCellInfo);
+            dbg.Model.CellsChanged += new Syncfusion.Windows.Forms.Grid.GridCellsChangedEventHandler(GridCellChanged);
             #endregion
         }
 
@@ -317,14 +364,56 @@ namespace CSV_Analyzer_Pro{
                 return false;
             }
         }
+
+        private void PromptUnsavedChangesCloseTab(TabExtraInfo tabInfo) {
+            UnsavedChangesSingleFile unsavedChangesBox = new UnsavedChangesSingleFile();
+            unsavedChangesBox.ShowDialog();
+
+            switch (unsavedChangesBox.userAnswer) {
+               case UnsavedChangesSingleFile.saveAndClose:
+                    Save(tabInfo.GetAssocaitedFileName(), tabControl1.SelectedIndex);
+                    DeleteTab();
+                    break;
+               case UnsavedChangesSingleFile.closeWithoutSaving:
+                    DeleteTab();
+                    break;
+               case UnsavedChangesSingleFile.Cancel:
+                    break;
+            }
+        }
+
+        private bool PromptUnsavedChangesOpenTab(TabExtraInfo tabInfo) {
+            UnsavedChangesSingleFile unsavedChangesBox = new UnsavedChangesSingleFile();
+            unsavedChangesBox.ShowDialog();
+
+            switch (unsavedChangesBox.userAnswer) {
+               case UnsavedChangesSingleFile.saveAndClose:
+                    Save(tabInfo.GetAssocaitedFileName(), tabControl1.SelectedIndex);
+                    return true;
+               case UnsavedChangesSingleFile.closeWithoutSaving:
+                    return true;
+               case UnsavedChangesSingleFile.Cancel:
+                    return false;
+               default:
+                    return false; // Should be unreachable
+            }
+        }
         #endregion
 
         #region Exiting Functions
         protected override void OnFormClosing(FormClosingEventArgs e) {
             base.OnFormClosing(e);
 
-            if (e.CloseReason == CloseReason.WindowsShutDown) return;
-            if (!_exiting) {
+            bool hasUnsavedChanges = false;
+
+            foreach (TabExtraInfo tabInfo in tabMetadataList) { 
+                if (tabInfo.QueryHasUnsavedChanges()) {
+                    hasUnsavedChanges = true;
+                }
+            }
+
+            if (e.CloseReason == CloseReason.WindowsShutDown) { return; }
+            if (!_exiting && !hasUnsavedChanges) {
                 switch (MessageBox.Show(this, "Are you sure you want to exit?", "Closing", MessageBoxButtons.YesNo)) {
                     case DialogResult.No:
                         e.Cancel = true;
@@ -335,6 +424,28 @@ namespace CSV_Analyzer_Pro{
                     default:
                         break;
                 }
+            }
+            else if (!_exiting && hasUnsavedChanges) {
+                PromptUnsavedChangesExit(e);
+            }   
+        }
+
+        private void PromptUnsavedChangesExit(FormClosingEventArgs e) {
+            UnsavedChangesMultipleFiles unsavedChangesBox = new UnsavedChangesMultipleFiles();
+            unsavedChangesBox.ShowDialog();
+
+            switch (unsavedChangesBox.userAnswer)
+            {
+                case UnsavedChangesMultipleFiles.saveAllAndClose:
+                    SaveAll();
+                    ExitAll();
+                    break;
+                case UnsavedChangesMultipleFiles.closeWithoutSaving:
+                    ExitAll();
+                    break;
+                case UnsavedChangesMultipleFiles.Cancel:
+                    e.Cancel = true;
+                    break;
             }
         }
 
@@ -360,15 +471,29 @@ namespace CSV_Analyzer_Pro{
             //card.WireGrid(dbg);
             DataTable dt = new DataTable();
 
-            tb.Text = "New..";
+            tb.Text = "New";
 
             tb.Controls.Add(dbg);
             tabControl1.TabPages.Add(tb);
             tabControl1.SelectedTab = tb;
+            
+            TabExtraInfo newTab = new TabExtraInfo(tabControl1.TabCount-1, "N/A");
+            tabMetadataList.Insert(tabControl1.TabCount-1, newTab);
         }
 
-        private void Save(string filePath) {
-            int index = tabControl1.SelectedIndex;
+        private void Save(string filePath, int index=-1) {
+            if (index == -1) {
+                index = tabControl1.SelectedIndex;
+                // I found that the usage of SelectedIndex in the Save functions resticted them to the current tab, which was a problem when trying to implement SaveAll.
+                // This should allow saving of data from any tab without interfering with current functions by defaulting to SelectedIndex if no other index is specified.
+            }
+
+            if (filePath == "N/A") {
+                return; // Skipping tabs with associatedFileName of N/A, mainly the Welcome page
+            }
+
+            TabExtraInfo tabInfo;
+            tabInfo = tabMetadataList.ElementAt(index);
 
             StringBuilder sb = new StringBuilder();
 
@@ -381,12 +506,20 @@ namespace CSV_Analyzer_Pro{
                 sb.AppendLine(string.Join(",", fields));
             }
 
+            bool success;
+
             try {
                 File.WriteAllText(filePath, sb.ToString());
+                success = true;
             } catch (Exception e) {
                 Console.WriteLine("An Exception occured when trying to save file");
+                success = false;
             }
 
+            if (success && tabInfo.QueryHasUnsavedChanges()) { // Avoiding removal of unsaved changes marker where it does not exist or where save has failed
+                tabInfo.SetHasUnsavedChanges(false);
+                tabControl1.TabPages[index].Text = tabControl1.TabPages[index].Text.Substring(1); // Return substring without unsaved changed marker
+            }
         }
 
         private void SaveAs() {
@@ -402,6 +535,11 @@ namespace CSV_Analyzer_Pro{
                 savePath = saveAs.FileName;
             }
 
+            TabExtraInfo tabInfo;
+            tabInfo = tabMetadataList.ElementAt(tabControl1.SelectedIndex);
+            tabInfo.SetAssociatedFileName(savePath);
+            tabControl1.SelectedTab.Text = savePath;
+
             Thread th = new Thread(() => Save(savePath));
             th.Start();
 
@@ -409,7 +547,17 @@ namespace CSV_Analyzer_Pro{
             //Save(savePath);
         }
 
+        private void SaveAll() {
+            int i;
+            TabExtraInfo tabInfo;
+            for (i = 0; i < tabControl1.TabCount; i++) {
+                tabInfo = tabMetadataList.ElementAt(i);
+                Save(tabInfo.GetAssocaitedFileName(), i);
+            }
+        }
+
         private void DeleteTab() {
+            tabMetadataList.RemoveAt(tabControl1.SelectedIndex);
             Debug.WriteLine("Delete Tab Called");
             int index = tabControl1.SelectedIndex;
             tabControl1.TabPages.RemoveAt(index);
